fix: port and delete api

+++ backend/controllers/product.controller.js
@@ -47,6 +47,10 @@ export const getProducts = async(req,res) => {
 
   export const deleteProduct = async(req,res) => {
     const {id} = req.params
+
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: "Invalid ID" });
+    }
   
     try {
       await Product.findByIdAndDelete(id);
